03.12.2020

+++ vrl/Zusatzaufgaben_Basics/Zusatzaufgaben.java
@@ -7,6 +7,6 @@ public class Zusatzaufgaben {
         System.out.print("Das sind");
         System.out.println(" Zusatzaufgaben");
         // 3
-        int a = (int) (17/2);
+        int a = 17.0/2.0; 
     }
 }
